Fix problem with not adding scenario chart
Checking parameter it for chart equality is not enough. There can be a
situation when real measurements and measurements from scenario are
added to the same chart. As a result `getParameterId` method will return
the same values, but series are totally different. The solution is to
use `getTimelineId` to check series equality.